Support Shielded VMs in Instance Template
Added support for `shielded_instance_config` as well as updated the default image project and family (already CentOS, just specified) to be the type of image that supports Shielded VMs. While I believe we should turn this on by default, since the images that support it are limited, we should leave it off by default to avoid breaking existing instances.